Added minttyrc
I'll sometimes use either cygwin or msys2 on a Windows machine if I have
to use Windows and I either a) don't have the Windows Subsystem for Linux
installed or b) am unable to use it for some reason. When I do use one
of the two (cygwin or msys2), this is the configuration I use for the
terminal. For everything else (such as running vim in cygwin/msys2), I
use the same dotfiles that I use everywhere else.