Added mbsync and vdirsyncer files
The checkmail script executes mbsync to sync local and remote mailboxes,
notmuch to update the mail database, and afew to perform some initial
tagging.

Next, the calendarsync script executes vdirsyncer to sync local and
remote calendars (Google Calendar in this case). At the moment, this is
all that the calendarsync script does, but it may do more in the future.

Both scripts are run automatically using systemd service timers. If I've
configured them right, these services shouldn't execute if a network
isn't available. If a network is available, the timers will activate
every five minutes to perform the sync operations.