Massive overhaul of init.vim, added ginit.vim
I'd list all the changes to init.vim here, but to be honest, I really
don't want to type that much.

Also, I may or may not have forgotten some of the change I made and
I'm not going to like, while I *could* do a diff, I really don't want
to.